backend/requirements.txt
fastapi
uvicorn
python-multipart
python-jose[cryptography]
passlib[bcrypt]
pydantic[email]
pymupdf
python-docx
beautifulsoup4
requests
playwright
python-dotenv
groq
ollama
pydantic-settings
openai
jinja2
supabase
python-docx
